pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["jinwoo <sds7629>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
apscheduler = "^3.10.1"
django-apscheduler = "^0.6.2"
django-cors-headers = "^4.2.0"
django-environ = "^0.10.0"
djangorestframework-simplejwt = "^5.2.2"
djangorestframework = "^3.14.0"
environ = "^1.0"
xmltodict = "^0.13.0"
requests = "^2.31.0"
pyjwt = "^2.8.0"
pillow = "^10.0.0"
drf-spectacular = "^0.26.4"
django-hitcount = "^1.3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
